+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added support for deploying Holoviz Panel applications with the `python-panel` content type.
+
 ### Fixed
 
 - When creating a Connect credential using token authentication or an API key